backend/requirements.txt
bcpandas==2.5.0
nltk==3.8.1
numpy==1.26.4
pandas==2.0.3
psycopg2==2.9.9
PyJWT==2.8.0
PyMySQL==1.1.0
pyodbc==5.0.1
SQLAlchemy==2.0.27
